requirements.txt
aiohttp_cors
beautifulsoup4
homeassistant-stubs
mypy
PyJWT
ruff
types-cffi
types-PyMySQL
types-pyRFC3339
types-python-dateutil
types-PyYAML
types-requests
types-setuptools
